backend/requirements.txt
Django==4.0.1
django-cors-headers
pymysql==1.0.2 # Move to Postgres?
djangorestframework
django-admin
django-environ
pymysql